+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #368 ECharts: support "itemStyle", "symbolSize" for "scatter" series #368 
 * #369 ECharts: support "itemStyle" for "pie" series
 * #370 ECharts: support "itemStyle", "symbolSize", "lineStyle" for "line" series
+* #371 ECharts: support "itemStyle" for "boxplot" series
 
 ## Release 1.0.0-RC2